Implement splatter partitioning for graph visualization
Refactored the main script to include a splatter partitioning function which performs breadth-first partitioning on a graph. This feature significantly enhances the visualization by dividing the graph into clearly distinguishable clusters. NetworkX is used to generate the graph and Matplotlib to visualize it. Also, added functionality for subsequent refinement of partitions.